CECS-327-proj/zero_mq/requirements.txt
pyzmq
fastapi (uvicorn)
